deduplicate buildSearchURLQuery
NOCHANGELOG
