+++ client/browser/src/shared/util/url.tsx
@@ -52,15 +52,3 @@ export function toAbsoluteBlobURL(ctx: AbsoluteRepoFile & Partial<PositionSpec> 
         ctx.position
     )}${toReferencesHash(ctx.referencesMode)}`
 }
-
-/**
- * Builds a URL query for given SearchOptions (without leading `?`)
- */
-export function buildSearchURLQuery(query: string): string {
-    const searchParams = new URLSearchParams()
-    searchParams.set('q', query)
-    return searchParams
-        .toString()
-        .replace(/%2F/g, '/')
-        .replace(/%3A/g, ':')
-}
